Add initial fetch logic to Root component
